+++ ssvm/loss_functions.py
@@ -27,8 +27,9 @@ import numpy as np
 
 from scipy.sparse import issparse
 
-from ssvm.kernel_utils import tanimoto_kernel
+from ssvm.kernel_utils import tanimoto_kernel_FAST as tanimoto_kernel
 from ssvm.kernel_utils import generalized_tanimoto_kernel_FAST as minmax_kernel
+from ssvm.kernel_utils import generalized_tanimoto_kernel_FAST as generalized_tanimoto_kernel
 
 
 def hamming_loss(y: np.ndarray, Y: np.ndarray) -> np.ndarray:
@@ -68,7 +69,7 @@ def tanimoto_loss(y: np.ndarray, Y: np.ndarray) -> np.ndarray:
     assert not issparse(y)
     assert not issparse(Y)
 
-    return 1 - tanimoto_kernel(np.atleast_2d(y), np.atleast_2d(Y), shallow_input_check=True).flatten()
+    return 1 - tanimoto_kernel(np.atleast_2d(y), np.atleast_2d(Y)).flatten()
 
 
 def minmax_loss(y: np.ndarray, Y: np.ndarray) -> np.ndarray:
@@ -83,3 +84,17 @@ def minmax_loss(y: np.ndarray, Y: np.ndarray) -> np.ndarray:
     assert not issparse(Y)
 
     return 1 - minmax_kernel(np.atleast_2d(y), np.atleast_2d(Y)).flatten()
+
+
+def generalized_tanimoto_loss(y: np.ndarray, Y: np.ndarray) -> np.ndarray:
+    """
+
+    :param y:
+    :param Y:
+    :return:
+    """
+    assert len(y.shape) == 1
+    assert not issparse(y)
+    assert not issparse(Y)
+
+    return 1 - generalized_tanimoto_kernel(np.atleast_2d(y), np.atleast_2d(Y)).flatten()